# Rework PriorityQueue for performance.