- Providing the comparer as a type argument that is a struct allows the calls to be devirtualised, leading to approx a 3x performance improvement.
- Use a single backing array, rather than a list of arrays.

----

Supersedes #20117, as requested. See https://github.com/OpenRA/OpenRA/pull/20117#issuecomment-1474392600 for original performance analysis.

|                                     Method |     Mean |   Error |  StdDev |
|--------------------------------- |---------:|--------:|--------:|
| `PriorityQueue<T, TComparer>` | 224.5 us | 2.00 us | 1.87 us |
|                   `PriorityQueue<T>` | 727.5 us | 4.74 us | 4.44 us |